Ajout de la méthode Parcelles Adjacentes dans la classe parcelle #1

+++ src/main/java/fr/cotedazur/univ/polytech/startingpoint/Joueur.java
@@ -29,8 +29,8 @@ public class Joueur {
 
     public Parcelle jouerParcelle(){
         Random rand = new Random();
-        Double x = rand.nextDouble();
-        Double y = rand.nextDouble();
+        int x = rand.nextInt();
+        int y = rand.nextInt();
         Position positionParcelle1 = new Position(x, y);
         Parcelle parcelle1 = new Parcelle(positionParcelle1);
         return parcelle1;
